fix stuff based on issue #10
