+++ kallsyms.c
@@ -1,46 +1,22 @@
 //https://github.com/h33p/kallsyms-lp
-#include "kallsyms.h"
-#include <linux/livepatch.h>
+#include <linux/module.h>
+#include <linux/completion.h>
+#include <linux/list.h>
+#include <linux/version.h>
 
-static struct klp_func funcs[] = {
-	{
-		.old_name = "kallsyms_lookup_name",
-		.new_func = kallsyms_lookup_name,
-	}, {}
-};
-
-static struct klp_func failfuncs[] = {
-	{
-		.old_name = "___________________",
-	}, {}
-};
-
-static struct klp_object objs[] = {
-	{
-		.funcs = funcs,
-	},
-	{
-		.name = "kallsyms_failing_name",
-		.funcs = failfuncs,
-	}, { }
-};
-
-static struct klp_patch patch = {
-	.mod = THIS_MODULE,
-	.objs = objs,
-};
-
-unsigned long kallsyms_lookup_name(const char *name)
-{
-	return ((unsigned long(*)(const char *))funcs->old_func)(name);
-}
-
-int init_kallsyms(void)
-{
-	int r = klp_enable_patch(&patch);
-
-	if (!r)
-		return -1;
-
+#if IS_ENABLED(CONFIG_KPROBES)
+#include "kallsyms_kp.c"
+#elif IS_ENABLED(CONFIG_LIVEPATCH)
+#include "kallsyms_lp.c"
+#elif LINUX_VERSION_CODE < KERNEL_VERSION(5,7,0)
+int init_kallsyms(void) {
 	return 0;
+}
+#else
+#error "No suitable kallsyms acquisition method!"
+#endif
+
+int kallsyms_lookup_fault(const char *sym) {
+	printk("kallsyms_lookup_name failed for: %s\n", sym);
+	return -EBADF;
 }
